+++ worlds/pokemon_bw/data/pokemon/types.py
@@ -1,0 +1,23 @@
+
+by_name: dict[str, int] = {
+    "Normal": 0,
+    "Fighting": 1,
+    "Flying": 2,
+    "Poison": 3,
+    "Ground": 4,
+    "Rock": 5,
+    "Bug": 6,
+    "Ghost": 7,
+    "Steel": 8,
+    "Fire": 9,
+    "Water": 10,
+    "Grass": 11,
+    "Electric": 12,
+    "Psychic": 13,
+    "Ice": 14,
+    "Dragon": 15,
+    "Dark": 16,
+    # "Fairy": 17,
+}
+
+by_id = {index: name for name, index in by_name.items()}