+++ worlds/pokemon_bw/generate/__init__.py
@@ -53,7 +53,7 @@ class SpeciesEntry:
     tm_hm_moves: TMHMMovesetData
     is_custom_form: bool
     custom_form_file: int
-    # bitflag, b0 = evolutions, b1 = plando evo override, b2 = base stats, b3 = catch rate
+    # bitflag, b0 = evolutions, b1 = plando evo override, b2 = base stats, b3 = catch rate, b4 = levelup moveset
     write: int = 0
     evo_line: EvoLine | None = None